move join to different controller
