+++ Server/API/rooms/join/join.go
@@ -1,0 +1,9 @@
+package join
+
+import (
+	"github.com/labstack/echo/v4"
+)
+
+func JoinController(e *echo.Echo) {
+	e.PATCH("/rooms/join/:roomId", patchJoin)
+}
