+++ Server/API/rooms/rooms.go
@@ -2,11 +2,13 @@ package rooms
 
 import (
 	"github.com/labstack/echo/v4"
+	"ishtaloo.io/API/rooms/join"
 )
 
 func RoomsController(e *echo.Echo) {
 	e.POST("/rooms", postRoom)
-	e.PATCH("/rooms/join/:roomId", patchJoin)
 	e.GET("/rooms", getRooms)
 	e.GET("/rooms/:roomId", getRoomById)
+
+	join.JoinController(e)
 }
